modify domain field "qty" data type

+++ grails-app/domain/foodpaint/Batch.groovy
@@ -43,7 +43,7 @@ class Batch  {
 	]
 
 	String name
-	long expectQty = 0//ERP無此資料
+	double expectQty = 0//ERP無此資料
 	Date dueDate //失效日
 
     /**
